Isolate main routine
The line 139 introduces a bug: it can rename a file that has not been looped
over yet, causing multiple other lines to fail bacause file no longer exists.

The main routine was moved to a function that is never called to preserve the
older code while new solutions are tested.